fixed performance issues
for some reason I have to use float64 when calculating delta time or it often reads 0, init joystick apparently fixes the stutters which I'll have to see on pc later and have to force game, also force the game to use discrete graphics on laptop

+++ SDL-Game/src/Game/Game.h
@@ -18,11 +18,11 @@ private:
 	void Tick();
 
 	bool m_bGameIsRunning;
-	FLOAT32 m_CurrentFrameTime;
-	FLOAT32 m_LastFrameTime;
-	FLOAT32 m_DeltaTime;
-	FLOAT32 m_FixedTickRate;
-	FLOAT32 m_FixedDeltaTime;
+	FLOAT64 m_CurrentFrameTime;
+	FLOAT64 m_LastFrameTime;
+	FLOAT64 m_DeltaTime;
+	FLOAT64 m_FixedTickRate;
+	FLOAT64 m_FixedDeltaTime;
 
 	GameRenderer m_GameRenderer;
 	EventManager m_EventManager;
